requirements.txt
crewai     
fastapi
uvicorn
python-dotenv
redis
pydantic
